+++ src/main/java/com/unzer/payment/models/ShippingTransactionData.java
@@ -15,6 +15,8 @@
  */
 package com.unzer.payment.models;
 
+import java.util.Objects;
+
 public class ShippingTransactionData {
     private String deliveryTrackingId;
     private String deliveryService;
@@ -62,5 +64,27 @@ public class ShippingTransactionData {
     public ShippingTransactionData setReturnTrackingId(String returnTrackingId) {
         this.returnTrackingId = returnTrackingId;
         return this;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        ShippingTransactionData that = (ShippingTransactionData) o;
+
+        if (!Objects.equals(deliveryTrackingId, that.deliveryTrackingId))
+            return false;
+        if (!Objects.equals(deliveryService, that.deliveryService))
+            return false;
+        return Objects.equals(returnTrackingId, that.returnTrackingId);
+    }
+
+    @Override
+    public int hashCode() {
+        int result = deliveryTrackingId != null ? deliveryTrackingId.hashCode() : 0;
+        result = 31 * result + (deliveryService != null ? deliveryService.hashCode() : 0);
+        result = 31 * result + (returnTrackingId != null ? returnTrackingId.hashCode() : 0);
+        return result;
     }
 }